[feature/getHome] 홈화면 조회 API

+++ src/main/java/com/hackathon/unid/review/domain/entity/Reviews.java
@@ -30,5 +30,5 @@ public class Reviews extends BaseEntity {
     @Column(nullable = false)
     private String reviewImage;
 
-    private Integer totalRating;
+    private Double totalRating;
 }
